requirements.txt
biopython
pandas
requests 